Use fallback-x11
On a Wayland session, this app does not need the X11 socket.

Using fallback-x11 causes Flatpak to not expose the X11 socket on a
Wayland session; and has the knock-on effect of not marking the app as
“(Potentially) Unsafe” on the Flathub website and in GNOME Software.

Specifying both fallback-x11 and x11 causes x11 to be ignored on
versions of Flatpak that support both. However, we can instead decide
that Flatpak 0.11.1 (Feb 2018), which introduced fallback-x11, is
available everywhere and drop x11.